requirements.txt
dateparser==1.3.*
discord.py==2.7.*
graphviz==0.21.*
Japanera==2.1.*
mcrcon==0.7.*
moviepy==2.2.*
nltk==3.9.*
numpy==2.4.*
opencv-python==4.13.*
pillow>=9.2,<12.0
psutil==7.2.*
psycopg2-binary==2.9.*
pypandoc==1.16.*
pyyaml==6.0.*
regex==2026.2.28
requests==2.32.*
tzdata  # This only needed if run on windows. I haven't set a version either to ensure it's up to date.
wordcloud==1.9.*
yt-dlp==2026.3.13
